Added list history, fixed negative numbers.
Added a class containing a BigDecimal and operatorFlag and made a list
of them to help make a history of calculations and to fix a problem with
multiplications involving negative numbers. Added a boolean that also
helps with the negative multiplications. Changed the way enumerators
worked, because my switch statement was flawed.