Added code for operators.
Added code for most operators, now they set the flags of the numbers,
still need to test to see if it's doing it right. Added the ( and )
operator symbols for the brackets, made the long press 7 and 8 buttons
control the bracket. In a quick test the brackets didn't show, need to
do better testing. Long calculations still don't work, it just works out
the last calculation entered. Amended the equals method to include else
if statements instead of just if statements.