Plus works with long equations.
Plus works with long equations on the condition that you only use plus.
The problem was that when i iterate through my for statement, testing
each number in the equation, I then remove a number, and increment the j
value of the for statement, meaning when I equate the objects.get(j) to
the answer objects.get(j) is never used again, objects.get(j+1) is used
but if i use that i don't think my overall method will work when i come
to do brackets. I think i need to rethink my for statement, instead of
testing a number and seeing what operator flag it has, I need to iterate
through each number and check which has the highest operator, a simple
mistake which i would have spotted earlier if I had my mind on long
equations from the start.