Added more operators to the new equals method
Added the other operators to the equals method. There are no glitches
just logic errors apparently, I get the wrong answer for some
calculations. The remove method forces the application to close.
hardcoding numbers for the for statement within it works but I haven't
gotten to the bottom of why my method doesn't work. Really need to work
on this for more than 30 minutes at a time :/