Fixed multiplying by negative numbers.
There was a little glitch were performing 3*-4 would equal -1 rather
than -12, this was because the subtract method would call the number's
flag to plus, overwriting the multiply. Added an if statement to the
subtract method to prevent this, pluses will overwrite multiply flags,
but subtracts wont. Which is the correct behaviour. Attempted bracket
implementation, doesn't quite work.