Fixed brackets
The brackets now work as they should, as far as I know, need to do
further testing. We increment a variable currentBracket when we open a
bracket, all numbers' bracketNo variable is set to currentBracket, when
we close the bracket the currentBracket variable is decremented. The
calculations in my equals method go in BIDMAS order only if the number's
bracketNo matches the highest bracket number. This way, if a bracket
inside a bracket is considered a deep bracket, the deepest brackets are
considered first as they have the highest bracket number.