Fixed subtract
I noticed subtract had stopped working. This was because I recently
editing how the plus method worked. Adding some of the code of the plus
method to the button that calls the plus method (which I did so I could
use the plus flag for subtract and plus), I had to do the same for
subtract.