Fixed subtract again
During testing I noticed 4-2^2 = 8 instead of 0. This was because my
implementation of subtraction and negative numbers was not intended for
long equations. pressing - in the above equation had set the 2 to -2,
meaning when it was squared (^2) the negative disappeared. I fixed this
by changing the subtraction method, now it essentially does this
4+-1*2^2. We add a number to history, -1 and give it the multiply flag,
we add it before we add our user inputted number and rely on the BIDMAS
in the equals method to sort out the rest. I still need to prune a few
things from my previous subtraction implementation.