Fixed subtract for exponents and divide
My method of just multiplying the next number by -1 when pressing
subtract didn't work as it should in some cases involving operators of
higher BIDMAS order, such as 2^-2 essentially turned into 2^-1*2 or even
2-1*2 and therefore gave the wrong answer. I added a condition to
subtract so that if the previous number has a power or divide flag it
sets the next number entered to negative immediately, instead of letting
it be set to negative in the equals method when multiply is called,
which would be too late. A new problem seems to have occurred where the
entering a bracket after a subtract replaces the subtract. Look into
that.