Fixed divide by zero, decimals and added radians.
Dividing by zero crashed the app, I made a simple clause to stop that
happening. If the first number entered used a decimal the app would
crash, i fixed this. It crashed because the decimal method checks to see
if the last number entered used a power sign, since we can't use decimal
exponents, when the first number is a decimal it tries to check outside
the list of numbers causing a crash. I also added a radians button, to
toggle between degrees and radians, it would be useful to change the
text on the button to display if we are using degrees or radians.