Fixed error when operator is input before number
I "fixed" this by simply ensuring that the object list that stores the
numbers is never empty, before it was only empty initially and when you
press clear, now I add a number (0) to the list initially and at the end
of the clear method. This fix also makes a previous fix for the
decimalMode method redundant, in both errors the problem was trying to
access a member of an empty list.