Attempted to fix visual glitches
I want open brackets to appear on the right on the number when you press
them, but then when the next number, the number that will be inside the
bracket, is entered the open bracket should appear on the left. While
trying to get this to work, I realised I had to fix another problem
first. When you press two operators the method used to update the
display is called after the number variable is set to 0, meaning the
number on screen is confusingly set to 0. To fix this I simply made a
new variable called displayNumber and set it's value to number, and
don't reset it to 0 when pressing operators then I use this new variable
to update the display. Pretty simple, I might need to add displayNumber
= number; to my pi() and trig functions? I also set my operator chars to
'\0' instead of ' '.  Otherwise in my display ,which is basically a
string like bracketChar + numberBigDecimal + operatorChar+ inverseChar +
HyperbolicChar; , if you used hyperbolic operator the h symbol would
appear two spaces away from the number, ' ' is a space '\0' is null
meaning no space. This will matter less when I make a new method to
allow multiple operators present on screen.