Allowed multiple operators in display
Allowed multiple operators to be displayed on screen. Calculations such
as 1*-3 would previously only display one operator at a time which is
potentially confusing. Changed the char operator to a String. Attempting
to add logic that correctly positions the brackets, when first opening a
bracket it should appear to the right of the last number entered, when
we enter another number it should go to the left of the new number,
entering another number ill see it disappear. Still some bugs I need to
iron out with that. There was also a problem with entering long
calculations, the number would change to 0, this was because I set the
value of number to zero before displaying it because of the operators
that were being called, I added a new BigDecimal called displayNumber
and updated it's value to the same as number whenever that was changed,
but I don't set it to zero at the same time number needs to be so we can
keep the value for display purposes. I also entered a error message when
dividing by zero.