Fixed divide by zero message
Last commit contained an update to the divide method that didn't work as
intended. The display should have been updated with an error message,
the problem was that the display was updated again with something else
that replaced the error message before the user has a chance to see it.